Change end-point for semantics
